email: Move a couple of useful functions to email_helpers.py.
